move to a cmake model
