+++ .gitignore
@@ -1,3 +1,6 @@
+bin
+CMakeFiles
+CMakeCache.txt
 *.a
 a.out
 *.o
