+++ spiner/sp5.hpp
