--- test/Makefile.kokkos
@@ -1,78 +1,0 @@
-# © (or copyright) 2019-2021. Triad National Security, LLC. All rights
-# reserved.  This program was produced under U.S. Government contract
-# 89233218CNA000001 for Los Alamos National Laboratory (LANL), which is
-# operated by Triad National Security, LLC for the U.S.  Department of
-# Energy/National Nuclear Security Administration. All rights in the
-# program are reserved by Triad National Security, LLC, and the
-# U.S. Department of Energy/National Nuclear Security
-# Administration. The Government is granted for itself and others acting
-# on its behalf a nonexclusive, paid-up, irrevocable worldwide license
-# in this material to reproduce, prepare derivative works, distribute
-# copies to the public, perform publicly and display publicly, and to
-# permit others to do so.
-
-KOKKOS_PATH = ${HOME}/kokkos
-SRC = test.cpp
-#SRC = convergence.cpp
-vpath %.cpp $(sort $(dir $(SRC)))
-
-KOKKOS_CXX_STANDARD=c++14
-
-default: build
-	echo "Start Build"
-
-#H5_DIR=/usr/local/hdf5-parallel
-#H5_INCLUDE=-I${H5_DIR}/include
-#H5_LIB=-L${H5_DIR}/lib
-#H5_LINK=-lhdf5_hl -lhdf5 -Wl,-rpath=${H5_DIR}/lib/
-
-ifneq (,$(findstring Cuda,$(KOKKOS_DEVICES)))
-CXX = ${KOKKOS_PATH}/bin/nvcc_wrapper
-CXXFLAGS = -O0 --expt-relaxed-constexpr -g
-CXXFLAGS += -DPORTABILITY_STRATEGY_KOKKOS #-DSPINER_USE_HDF
-LINK = ${CXX}
-#LDFLAGS = 
-#EXE = simple_test_cuda
-KOKKOS_DEVICES = "Cuda"
-KOKKOS_ARCH = "Volta70"
-KOKKOS_CUDA_OPTIONS += "enable_lambda;rdc"
-#CXXFLAGS = -O3 --expt-relaxed-constexpr #-x cuda --cuda-gpu-arch=sm_37 --stdlib=libstdc++ -lineinfo
-LINK = ${CXX}
-LDFLAGS =
-EXE = simple_test_cuda
-else
-CXX = g++
-CXXFLAGS = -O0 -g -DPORTABILITY_STRATEGY_KOKKOS #-DSPINER_USE_HDF
-LINK = ${CXX}
-LDFLAGS = #-lhdf5_hl -lhdf5 -Wl,-rpath=${H5_DIR}/lib/
-EXE = simple_test
-#EXE = convergence_test
-KOKKOS_DEVICES = "OpenMP"
-KOKKOS_ARCH = ""
-#KOKKOS_DEBUG=yes
-endif
-
-EXTRA_INC += -I../Catch2/single_include/catch2 -I../ports-of-call -I../
-#EXTRA_INC += ${H5_INCLUDE}
-DEPFLAGS = -M
-
-OBJ = $(notdir $(SRC:.cpp=.o))
-#LIB = -L${H5_DIR}/lib
-
-include $(KOKKOS_PATH)/Makefile.kokkos
-
-build: $(EXE)
-
-test: $(EXE)
-	./$(EXE)
-
-$(EXE): $(OBJ) $(KOKKOS_LINK_DEPENDS)
-	$(LINK) $(KOKKOS_LDFLAGS) $(LDFLAGS) $(EXTRA_PATH) $(OBJ) $(KOKKOS_LIBS) $(LIB) -o $(EXE)
-
-clean: kokkos-clean 
-	rm -f *.o *.cuda *.host *.rocm ${EXE}
-
-# Compilation rules
-
-%.o:%.cpp $(KOKKOS_CPP_DEPENDS)
-	$(CXX) $(KOKKOS_CPPFLAGS) $(KOKKOS_CXXFLAGS) $(CXXFLAGS) $(EXTRA_INC) -c $< -o $(notdir $@)
